added check to see if chrome-win32 exists before deleting existing chrome-win32.

+++ updatechromium.sh
@@ -11,18 +11,25 @@ echo "done!"
 echo "http://commondatastorage.googleapis.com/chromium-browser-snapshots/Win/$LATEST/chrome-win32.zip"
 echo -n "Downloading build for $LATEST... "
 wget http://commondatastorage.googleapis.com/chromium-browser-snapshots/Win/$LATEST/chrome-win32.zip
-echo "done!"
 
-echo -n "Deleting old install... "
-rm -rf chrome-win32
-echo "done!"
+if [ $? -ne 0 ]
+then 
+	echo "chrome-win32.zip not found, aborting.."
+else 
+	echo "done!"
+		
+	echo -n "Deleting old install... "
+	rm -rf chrome-win32
+	echo "done!"
 
-echo -n "Unpacking new install... "
-unzip -q chrome-win32.zip
-echo "done!"
+	echo -n "Unpacking new install... "
+	unzip -q chrome-win32.zip
+	echo "done!"
 
-echo -n "Cleaning up... "
-rm -f chrome-win32.zip
-echo "done!"
+	echo -n "Cleaning up... "
+	rm -f chrome-win32.zip
+	echo "done!"
+fi
+
 
 exit 0
